+++ symmetria/__init__.py
@@ -4,7 +4,7 @@ from symmetria.generators.random.api import random, random_generator
 from symmetria.generators.algorithm.api import generate
 from symmetria.elements.cycle_decomposition import CycleDecomposition
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 __all__ = [
     "__version__",
     "generate",